demo/requirements.txt
numpy
scipy
pyyaml
matplotlib
plotly
nbformat>=4.2.0
seaborn
pandas